Stats_discord_bot/requirements.txt
discord.py==2.3.2
pilmoji==0.1.8
aiohttp==3.8.5
Pillow==10.1.0
emoji==1.7.0
numpy==1.26.4
matplotlib==3.8.1
python-dotenv==1.0.0
pytz==2025.7
cryptography==41.0.4
redis==5.4.1
asyncpg==1.2.1

